Do not restore the generator context in internal `yield*` steps.
This PR changes the behavior of `yield*` so that the context inside
the generator is captured before yielding the first time, and only
restored when `yield*` ends (rather than for every call to the inner
iterator).

This means that the context is still lexically preserved within the
generator body, but `yield*` of an iterator behaves as if the generator's
caller was directly iterating the inner iterator: if the inner iterator
is not a generator but a plain handwritten iterator, it will have access
to the context of `.next()`'s caller.

This means that, using `yield*`, generators can manually access variables
from the `.next()` caller without needing a wrapper around the whole
generator. For example:
```javascript

function* gen() {
  console.log(myVar.get()); // logs 1

  const [value, outerVar] = yield* readAsyncVariable(myVar, "hi");

  console.log(myVar.get(), outerVar, value); // logs 1, 2, "there"
}

let it = myVar.run(1, () => gen());
let res = it.next(); // logs 1, returns "hi"
myVar.run(2, () => it.next("there")); // logs 1, 2, "there"

function readAsyncVariable(asyncVar, yieldValue) {
  let done = false;
  return {
    [Symbol.iterator]: () => ({
      next(v) {
        if (!done) {
          done = true;
          return { value: yieldValue, done: false };
        }
        return { value: [v, asyncVar.get()], done: true };
      }
    });
  }
}
```